Move makeRequest to helper fn
